refactor(server): isolate hosted mirror pairing
